// src/generated/models/Pulse_favoritesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pulse_favoritesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pulse_favoritesstatecode = keyof typeof Pulse_favoritesstatecode;
export const Pulse_favoritesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pulse_favoritesstatuscode = keyof typeof Pulse_favoritesstatuscode;

export interface Pulse_favoritesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  "pulse_App@odata.bind"?: string;
  pulse_favoriteid: string;
  pulse_name?: string;
  pulse_order?: number;
  "pulse_user@odata.bind"?: string;
  statecode: Pulse_favoritesstatecode;
  statuscode?: Pulse_favoritesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pulse_favorites extends Pulse_favoritesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pulse_appname?: string;
  pulse_username?: string;
  pulse_useryominame?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pulse_app?: object;
  _pulse_app_value?: string;
  pulse_user?: object;
  _pulse_user_value?: string;
}